chore: Remove use of unsupported callbacks
Connection callbacks are yet to be introduced in the new AMQP SDK,
and 'onDeviceUnregistered' is no longer required (it was, during a
transition phase in which KNoT arquitecture temporarily had no FOG
-- now it does), remove their usage so they're never called from
within knot-cloud-connector.

However, keep their declarations in the constructor, as the base
connector (CESARBR/knot-fog-connector) will still call these callbacks
(they will do nothing, this time). This avoids breaking the application.