refactor: Move callbacks to the constructor
Depending on how the client behaves, having the callbacks declared
in start() will result in side effects. As a result, the base connector
(CESARBR/knot-fog-connector) will not set the callbacks in time and upon
calling a callback, a `<callback> is not a function` error will be
returned.

Move all callback declarations to the constructor to avoid the mentioned
side effects.